+++ frontend/components/Canvas.jsx
