+++ frontend/components/Canvas.styl
