+++ frontend/components/CanvasContainer.jsx
