+++ frontend/components/Carousel.jsx
