+++ frontend/components/Carousel.styl
