+++ frontend/components/CarouselContainer.jsx
