+++ frontend/components/LoadingIndicator.jsx
