+++ frontend/components/LoadingIndicator.styl
