+++ frontend/redux/reducers/__tests__/loading.test.js
