+++ frontend/services/fabric.js
