+++ frontend/utils/deep-freeze.js
