+++ package.json
@@ -10,8 +10,13 @@
   ],
   "scripts": {
     "build": "webpack",
+    "migrate": "careen -c backend/db/migrations/careen",
+    "migrate:up": "careen -A -e -c backend/db/migrations/careen",
+    "migrate:down": "careen -R -e -c backend/db/migrations/careen",
+    "migrate:create": "careen -C -c backend/db/migrations/careen",
+    "migrate:reset": "careen -R -a -e -n 10000 -c backend/db/migrations/careen; npm run migrate:up",
     "start": "webpack-dev-server",
-    "lint": "eslint --ext js,jsx src",
+    "lint": "eslint --ext js,jsx frontend",
     "test": "node jest/run.js"
   },
   "jest": {
@@ -31,8 +36,12 @@
     "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(js|jsx)$"
   },
   "dependencies": {
+    "babel-cli": "^6.23.0",
     "babel-polyfill": "^6.23.0",
+    "careen": "^0.1.1",
+    "dotenv": "^4.0.0",
     "fabric": "^1.7.6",
+    "pg": "^6.1.2",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-hot-loader": "^3.0.0-beta.6",
